config bootstrap.yml v6

+++ trabajo/src/main/java/cloud/trabajo/service/UserServiceImplement.java
@@ -1,9 +1,12 @@
 package cloud.trabajo.service;
 
+import cloud.trabajo.client.BookingClient;
 import cloud.trabajo.entity.Users;
+import cloud.trabajo.model.Booking;
 import cloud.trabajo.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.GeneratedValue;
@@ -15,6 +18,9 @@ public class UserServiceImplement implements UserService{
 
 
     private final UserRepository userRepository;
+
+    @Autowired
+    private BookingClient bookingClient;
 
     @Override
     public List<Users> listAllUsers() {
@@ -34,11 +40,26 @@ public class UserServiceImplement implements UserService{
 
     @Override
     public Users deleteUser(Long id) {
-
+        Boolean b = false;
         Users usersDB = getUser(id);
         if (null == usersDB){
             return null;
         }
+
+        List<Booking> bookings = bookingClient.listBooking().getBody();
+        for (int i = 0; i<bookings.size(); i++) {
+
+            if(bookings.get(i).getUserid()==usersDB.getId()){
+                System.out.println("El usuario no puede ser eliminado porque tiene pendiente una reserva");
+                b=true;
+            }
+
+        }
+
+        if(b==true){
+            return null;
+        }
+
         userRepository.deleteById(id);
         return usersDB;
     }
